+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All Notable changes to `ElementFinder` will be documented in this file
+
 ## 0.2.1 [Unreleased]
+
+### Changed  
+- #92 Require second parameter `StringCollection::replace`
+
 
 ## 0.2.0 [2017-11-02]